requirements.txt
atomicwrites==1.4.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.2.1
dj-database-url==0.5.0
Django==2.2.15
django-bootstrap-form==3.4
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jsonpath==0.82
Markdown==3.2.2
model-mommy==2.0.0
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
